requirements.txt
# Core dependencies
pytest==7.4.4
# Gymnasium with Atari and ROM auto-acceptance (installs AutoROM and ROMs)
gymnasium[atari,accept-rom-license]==0.29.1
# torch==2.7.0
# torchvision==0.22.0
pillow==11.2.1
numpy==1.24.3
opencv-python==4.9.0.80
wandb==0.16.3
# For progress bars
tqdm==4.67.1
matplotlib==3.10.3
imageio==2.37.0
imageio[ffmpeg]
pygame==2.6.1
# Note:
# After installing requirements, run the following command to ensure all Atari ROMs are available:
#   python -m AutoROM --accept-license
# This will download and install all required ROMs for Atari environments.
# If you encounter ROM errors, re-run the above command or see part1.md for troubleshooting.

ale-py==0.8.1
AutoROM.accept-rom-license  # This automatically accepts Atari ROM license 